[#4851] Default to IPv4 addresses only for resolution and local addr selection
Summary:
In https://github.com/yugabyte/yugabyte-db/commit/5a88e2fe053180ae87d2c6fff061244ccac9c45f, a getaddrinfo call was changed to use AF_UNSPEC instead of AF_INET to return both IPv4 and IPv6 results for resolution. This means that localhost now resolves to both 127.0.0.1 and [::1] on most modern systems that have these entries in /etc/hosts.

However, in https://github.com/yugabyte/yugabyte-db/runs/787021393, it seems there is a github workflow host that resolves localhost to [::1] but has an error binding to [::1].

The issue doesn't happen on a vanilla Ubuntu 18.04 GCP host with a similar IPv6 entry added for localhost. I'm not sure how many hosts exist with this issue but to be safe, I am restricting DNS resolution and local address seletion to only use IPv4 hosts. This behavior should be roughly similar to earlier.

Test Plan: Ran this fix through the workflow as part of a PR and verified that the create snapshot failure no longer occurs.

Reviewers: bogdan, sergei, mikhail

Reviewed By: mikhail

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D8734